feat: Split code in service structure. fix: Avoid division by 0. feat: Add requirements.txt
