+++ requirements.txt
@@ -1,0 +1,2 @@
+matplotlib~=3.2.1
+Pillow~=7.1.2
