+++ src/Constants.py
@@ -1,0 +1,21 @@
+import logging
+
+# Folders.
+FOLDER_NAME_OUTPUT_JPG = 'jpg'
+FOLDER_NAME_OUTPUT_RESULT = 'selected_{}'
+
+# Formats.
+IMAGE_FORMAT_TIF = 'TIF'
+IMAGE_FORMAT_TIFF = 'TIFF'
+IMAGE_FORMAT_BMP = 'BMP'
+IMAGE_FORMAT_PNG = 'PNG'
+IMAGE_FORMAT_JPEG = 'JPEG'
+IMAGE_FORMAT_JPG = 'JPG'
+
+# Tiers.
+IMAGE_INTERVAL_TIERS = 10
+
+# Lists.
+ACCEPTED_IMAGE_FORMATS = [IMAGE_FORMAT_TIF, IMAGE_FORMAT_TIFF, IMAGE_FORMAT_BMP, IMAGE_FORMAT_PNG]
+LOGGING_LEVELS = [logging.ERROR, logging.WARNING, logging.INFO, logging.DEBUG]
+PERCENT_TIERS = dict((i, 0) for i in range(0, 100, IMAGE_INTERVAL_TIERS))
